requirements.txt
# Core AI/Agent
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Database
duckdb>=1.0.0
sqlalchemy>=2.0.0

# Data Generation & APIs
faker>=30.0.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0

# Visualization
plotly>=5.24.0
matplotlib>=3.9.0

# Forecasting
prophet>=1.1.5
scikit-learn>=1.5.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Development
pytest>=8.0.0
black>=24.0.0
